Rename createAuthUser file
