+++ src/__tests__/withAuthUserTokenSSR.test.ts
@@ -2,7 +2,7 @@ import { GetServerSidePropsContext } from 'next'
 import { setConfig } from 'src/config'
 import getMockConfig from 'src/testHelpers/createMockConfig'
 import { createMockFirebaseUserAdminSDK } from 'src/testHelpers/authUserInputs'
-import createAuthUser from 'src/createAuthUser'
+import createAuthUser from 'src/createUser'
 import createMockNextContext from 'src/testHelpers/createMockNextContext'
 import { AuthAction } from 'src/AuthAction'
 import getUserFromCookies from 'src/getUserFromCookies'
